backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
aiofiles==23.2.1

# LLM Providers — only the one matching LLM_PROVIDER in .env is actually called
groq==0.9.0
google-generativeai==0.7.2
anthropic==0.40.0
openpyxl==3.1.2    
xlrd==2.0.1 